refactor (gravsearch): move constants to separate file
